refactor(HTML): clean up code & move out constants in a separate file
